protect project - untested
