+++ app/page.tsx
@@ -4,7 +4,6 @@ import AuthForm from "./auth-form"
 const Home: React.FC = () => {
   return (
     <>
-      <h1>Hello World</h1>
       <AuthForm />
     </>
   )
